+++ src/Observers/EeCategoryObserverTrait.php
@@ -12,7 +12,7 @@
  * PHP version 5
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
- * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/techdivision/import-product-ee
  * @link      http://www.techdivision.com
@@ -26,7 +26,7 @@ use TechDivision\Import\Category\Ee\Utils\MemberNames;
  * Observer that create's the category itself for the Magento 2 EE edition.
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
- * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/techdivision/import-product-ee
  * @link      http://www.techdivision.com
@@ -55,6 +55,19 @@ trait EeCategoryObserverTrait
     {
         $this->setLastEntityId($category[MemberNames::ENTITY_ID]);
         $this->setLastRowId($category[$this->getPkMemberName()]);
+    }
+
+    /**
+     * Return's the entity ID for the passed path.
+     *
+     * @param string $path The path to return the entity ID for
+     *
+     * @return integer The mapped entity ID
+     * @throws \Exception Is thrown, if the path can not be mapped
+     */
+    protected function mapPath($path)
+    {
+        return $this->getSubject()->mapPathRowId($path);
     }
 
     /**