+++ src/Observers/EeUrlRewriteObserver.php
@@ -1,0 +1,43 @@
+<?php
+
+/**
+ * TechDivision\Import\Category\Ee\Observers\EeUrlRewriteObserver
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-category-ee
+ * @link      http://www.techdivision.com
+ */
+
+namespace TechDivision\Import\Category\Ee\Observers;
+
+use TechDivision\Import\Category\Observers\UrlRewriteObserver;
+
+/**
+ * Observer that creates/updates the category's URL rewrites for Magento 2 EE.
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-category-ee
+ * @link      http://www.techdivision.com
+ */
+class EeUrlRewriteObserver extends UrlRewriteObserver
+{
+
+    /**
+     * The trait providing category import functionality.
+     *
+     * @var \TechDivision\Import\Category\Ee\Observers\EeCategoryObserverTrait
+     */
+    use EeCategoryObserverTrait;
+}