+++ src/Subjects/EeBunchSubject.php
@@ -12,7 +12,7 @@
  * PHP version 5
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
- * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/techdivision/import-category-ee
  * @link      http://www.techdivision.com
@@ -22,12 +22,13 @@ namespace TechDivision\Import\Category\Ee\Subjects;
 
 use TechDivision\Import\Utils\RegistryKeys;
 use TechDivision\Import\Category\Subjects\BunchSubject;
+use TechDivision\Import\Category\Ee\Utils\MemberNames;
 
 /**
  * A SLSB that handles the process to import category bunches.
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
- * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/techdivision/import-category-ee
  * @link      http://www.techdivision.com
@@ -50,23 +51,22 @@ class EeBunchSubject extends BunchSubject
     protected $pathRowIdMapping = array();
 
     /**
-     * Clean up the global data after importing the bunch.
+     * Intializes the previously loaded global data for exactly one bunch.
      *
      * @param string $serial The serial of the actual import
      *
      * @return void
      */
-    public function tearDown($serial)
+    public function setUp($serial)
     {
 
-        // call parent method
-        parent::tearDown($serial);
+        // load the available category path => row ID mappings
+        foreach ($this->getCategoryBunchProcessor()->getCategoriesWithResolvedPath() as $resolvedPath => $category) {
+            $this->pathRowIdMapping[$resolvedPath] = $category[MemberNames::ROW_ID];
+        }
 
-        // load the registry processor
-        $registryProcessor = $this->getRegistryProcessor();
-
-        // update the status up the actual import with SKU => row ID mapping
-        $registryProcessor->mergeAttributesRecursive($this->serial, array(RegistryKeys::PATH_ROW_ID_MAPPING => $this->pathRowIdMapping));
+        // prepare the callbacks
+        parent::setUp($serial);
     }
 
     /**
@@ -101,5 +101,47 @@ class EeBunchSubject extends BunchSubject
     public function addPathRowIdMapping($path)
     {
         $this->pathRowIdMapping[$path] = $this->getLastRowId();
+    }
+
+    /**
+     * Removes the mapping, e. g. when a category has been deleted.
+     *
+     * @param string $path The path to delete the mapping for
+     *
+     * @return void
+     */
+    public function removePathRowIdMapping($path)
+    {
+        unset($this->pathRowIdMapping[$path]);
+    }
+
+    /**
+     * Return's the entity ID for the passed path.
+     *
+     * @param string $path The path to return the entity ID for
+     *
+     * @return integer The mapped entity ID
+     * @throws \Exception Is thrown, if the path can not be mapped
+     */
+    public function mapPathRowId($path)
+    {
+
+        // query whether or not a entity ID for the passed path has been mapped
+        if (isset($this->pathRowIdMapping[$path])) {
+            return $this->pathRowIdMapping[$path];
+        }
+
+        // throw an exception if not
+        throw new \Exception(sprintf('Can\'t map path %s to any row ID', $path));
+    }
+
+    /**
+     * Return's the PK column name to create the product => attribute relation.
+     *
+     * @return string The PK column name
+     */
+    protected function getPrimaryKeyMemberName()
+    {
+        return MemberNames::ROW_ID;
     }
 }